Fixes #5 where it could not extract the uploaderUrl due to Exception on Android 4.4.2
It seems that the Java SDK class URL' on android 4.4.2 is more strict if I
do not declare a protocol like https:// NewPipeExtractor
(in Utils.stringToURL(...)) tries to add https:// but in some cases it
will not be applied and crashes Therefore this crash happend there.

It also uncovered a bug where I was suspcection all images are from
i.rmbl.ws but there are also images from sp.rmbl.ws. So I changed it
to only check if the authority contains rmbl.ws