[Bitchute] Workaround missing duration extraction possibility in BitchuteStreamExtractor
The duration is not easily extractable without using JavaScript
at leas AFAIK. If you know a better way to extract the duration
let me know.

1. workaround -- get duration from that was populated by
BitchuteSearchExtractor, BitchuteTrendingStreamInfoItemExtractor,
BitchuteChannelStreamInfoItemExtractor and store it into a cache.

2. workaround -- search for the title on Bitchute and extract
duration from search result and store it into the cache

3. TODO workaround -- could be a fix in Player in NewPipe
that will extract the duration if possible from the video file